refactor: Rename github_action.rs to github_actions.rs
